+++ src/main/java/net/fyoncle/elysiumdaystweaks/utility/constants/Constants.java
@@ -7,6 +7,7 @@ public class Constants {
         public static final String MODRINTH_API_LINK = "https://api.modrinth.com/";
         public static final String ELYSIUM_DAYS_PAGE_LINK = "https://modrinth.com/modpack/elysium-days/versions";
         public static final String DISCORD_LINK = "https://discord.gg/WFpDr7zY8Z";
+        public static final String WIKI_LINK = "https://moddedmc.wiki/en/project/elysium-days/docs/getting_started";
         public static final String RAM_GUIDE_LINK = "https://moddedmc.wiki/en/project/elysium-days/docs/guides/memory_allocation";
     }